MAJOR [Added a mini-game and changed the way the COPY/PASTE works]
Added some text color to certain chat messages like errors, debugging,
etc.

Added a mini-game 'DoubleOrNothing'.
Also added a yml file into the jar thats is copied to the server when
it doesnt exist yet. That is used for the PASTE function.
So we can play the game.

Added GameHandling and EventHandling for games.

Improved the COPY/PASTE function more.
When PASTE is used with a game then we can get a entire list of all
the blocks that we can use for GameHandling and specific events.
Example: stop breaking game blocks by player or other players
or check which button in the game is pressed.

Removed unused code. Added extra comments and TODO's.