FEATURE [Experimental pasting function in a TaskTimer]
!! The current pasting is still active at this time !!
The experimental pasting function is added but NOT used yet.

What it does:
When pasting bigger portions of the world then you now will see it
get build.

This should also help with the server load.
If a COPY is to big and you'll try to paste it then the server could
crash.

I dont know how big a COPY can be without crashing.
I could COPY/PASTE an entire village, but the server crashed when I
tried to COPY/PASTE something that was 444 blocks in the distance,
85 blocks high and 526 blocks wide.