PATCH [Still a problem with pasting chunks]
Maybe this is not the correct approach to build something from a file.
Besides, this is not backwards compatible with the first CopyPaste.

Therefor I stop developing this function for now.
I think its better to focus on the normal copy/paste that does seem
to work now with a TaskTimer.